+++ packages/eas-cli/src/utils/profiles.ts
@@ -1,38 +1,45 @@
 import { Platform } from '@expo/eas-build-job';
-import { errors } from '@expo/eas-json';
+import { BuildProfile, EasJsonReader, SubmitProfile, errors } from '@expo/eas-json';
 
 import Log from '../log';
 
-export type ProfileData<T> = {
-  profile: T;
+type ProfileType = 'build' | 'submit';
+
+type EasProfile<T extends ProfileType> = T extends 'build'
+  ? BuildProfile<Platform>
+  : SubmitProfile<Platform>;
+
+export type ProfileData<T extends ProfileType> = {
+  profile: EasProfile<T>;
   platform: Platform;
   profileName: string;
 };
 
-export async function getProfilesAsync<T>({
+export async function getProfilesAsync<T extends ProfileType>({
+  projectDir,
   platforms,
   profileName: profileNameArg,
-  // eslint-disable-next-line async-protect/async-suffix
-  readProfileAsync,
+  type,
 }: {
+  projectDir: string;
   platforms: Platform[];
   profileName?: string | null;
-  readProfileAsync: (platform: Platform, profileName: string) => Promise<T>;
+  type: T;
 }): Promise<ProfileData<T>[]> {
   const results = platforms.map(async function (platform) {
-    let profile;
+    let profile: EasProfile<T>;
     let profileName = profileNameArg;
 
     if (!profileName) {
       try {
-        profile = await readProfileAsync(platform, 'production');
+        profile = await readProfileAsync({ projectDir, platform, type, profileName: 'production' });
         profileName = 'production';
       } catch (errorOuter) {
         if (errorOuter instanceof errors.InvalidEasJsonError) {
           throw errorOuter;
         }
         try {
-          profile = await readProfileAsync(platform, 'release');
+          profile = await readProfileAsync({ projectDir, platform, type, profileName: 'release' });
           profileName = 'release';
           Log.warn(
             'The default profile changed from "release" to "production". We detected that you still have a "release" build profile, so we are using it. Update eas.json to have a profile named "production" under the `build` key, or specify which profile you\'d like to use with the --profile flag. This fallback behavior will be removed in the next major version of EAS CLI.'
@@ -41,11 +48,11 @@ export async function getProfilesAsync<T>({
           if (errorInner instanceof errors.InvalidEasJsonError) {
             throw errorInner;
           }
-          throw new Error('There is no profile named "production" in eas.json');
+          throw new Error(`There is no ${type} profile named "production" in eas.json`);
         }
       }
     } else {
-      profile = await readProfileAsync(platform, profileName);
+      profile = await readProfileAsync({ projectDir, platform, type, profileName });
     }
 
     return {
@@ -56,4 +63,23 @@ export async function getProfilesAsync<T>({
   });
 
   return await Promise.all(results);
+}
+
+async function readProfileAsync<T extends ProfileType>({
+  projectDir,
+  platform,
+  type,
+  profileName,
+}: {
+  projectDir: string;
+  platform: Platform;
+  type: ProfileType;
+  profileName: string;
+}): Promise<EasProfile<T>> {
+  const easJsonReader = new EasJsonReader(projectDir);
+  if (type === 'build') {
+    return (await easJsonReader.readBuildProfileAsync(platform, profileName)) as EasProfile<T>;
+  } else {
+    return (await easJsonReader.readSubmitProfileAsync(platform, profileName)) as EasProfile<T>;
+  }
 }